femverd-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
cryptography
python-dotenv